+++ src/components/home/income/Income.svelte
@@ -1,69 +1,77 @@
 <script lang="ts">
     import MoneyCell from "$components/MoneyCell.svelte";
-    import { formatAmountWithoutSymbol, formatDate, getCurrencySymbol, type Income } from "$lib/lib";
-    import { useSelect } from "$lib/select.svelte";
+    import {
+        formatAmountWithoutSymbol,
+        formatDate,
+        getCurrencySymbol,
+        type Income,
+        type IncomeStream,
+        type Account,
+    } from "$lib/lib";
     import { appStore } from "$lib/state.svelte";
     import DatePicker from "../../DatePicker.svelte";
+    import SelectButton from "$components/SelectButton.svelte";
 
-	type Props = {
-		income: Income
-	}
+    type Props = {
+        income: Income;
+    };
 
-	const { income }: Props = $props();
+    const { income }: Props = $props();
     const symbol = getCurrencySymbol(income.currencyCode);
-    let formattedAmount = $derived.by(()=>formatAmountWithoutSymbol(income.amount,{ currency:income.currencyCode }));
+    let formattedAmount = $derived.by(() =>
+        formatAmountWithoutSymbol(income.amount, {
+            currency: income.currencyCode,
+        }),
+    );
 
-    const { select,options } = useSelect({
-        items: appStore.incomeStreams, // FIXME: income streams
-        toOption: (stream) => {return { value: stream.id, label: stream.title };},
-        onChange: ({ item }) => appStore.transactions.editIncome({ id: income.id,incomeStreamId: item.id }),
-    });
-
-    const { select: accountSelect,options: accountSelectOpts } = useSelect({
-        items: appStore.accounts,
-        toOption: (account) => {return { value: account.id, label: account.name };},
-        onChange: ({ item }) => appStore.transactions.editIncome({ id: income.id,accountId: item.id }),
-    });
-
-    function updateDate(year: number,month: number,day: number){
-        appStore.transactions.editIncome({ id: income.id, date: `${year}-${month}-${day}` });
+    function updateDate(year: number, month: number, day: number) {
+        appStore.transactions.editIncome({
+            id: income.id,
+            date: `${year}-${month}-${day}`,
+        });
     }
 
-    async function updateAmount(amount: string){
+    async function updateAmount(amount: string) {
         await appStore.transactions.editIncome({ id: income.id, amount });
+    }
+
+    async function editIncomeStream(item: IncomeStream) {
+        await appStore.transactions.editIncome({
+            id: income.id,
+            incomeStreamId: item.id,
+        });
+    }
+
+    async function editAccount(item: Account) {
+        await appStore.transactions.editIncome({
+            id: income.id,
+            accountId: item.id,
+        });
     }
 </script>
 
 <div class="data-cell flex justify-between items-center">
     <p>{income.incomeStream?.title ?? " "}</p>
-    <button {...select.trigger} class="icon-btn icon-btn-small icon-btn-neutral"><i class="ph ph-caret-down"></i></button>
-    <ul {...select.content} class="popup-overlay space-y-1 w-fit! overflow-auto">
-        {#each options as  option (option.value)}
-            {@const selected = select.isSelected(option)}
-            <li {...select.getOption(option)} data-selected={selected} class="menu-item w-fit min-w-[min-content]">
-                {option.label}
-            </li>
-        {/each}
-    </ul>
+    <SelectButton
+        items={appStore.incomeStreams}
+        toOption={(c) => {
+            return { label: c.title, value: c.id };
+        }}
+        onChange={editIncomeStream}
+    />
 </div>
 <div class="data-cell flex justify-between items-center">
     <p>{income.account?.name ?? " "}</p>
-    <button {...accountSelect.trigger} class="icon-btn icon-btn-small icon-btn-neutral"><i class="ph ph-caret-down"></i></button>
-    <ul {...accountSelect.content} class="popup-overlay space-y-1 w-fit! overflow-auto">
-        {#each accountSelectOpts as  option (option.value)}
-            {@const selected = accountSelect.isSelected(option)}
-            <li {...accountSelect.getOption(option)} data-selected={selected} class="menu-item w-fit min-w-[min-content]">
-                {option.label}
-            </li>
-        {/each}
-    </ul>
+    <SelectButton
+        items={appStore.accounts}
+        toOption={(c) => {
+            return { label: c.name, value: c.id };
+        }}
+        onChange={editAccount}
+    />
 </div>
 <li class="data-cell flex items-center justify-between">
-	<p>{formatDate(income.date)}</p>
-	<DatePicker onDateChange={updateDate}/>
+    <p>{formatDate(income.date)}</p>
+    <DatePicker onDateChange={updateDate} />
 </li>
-<MoneyCell
-    {symbol}
-    amount={formattedAmount}
-    onUpdate={updateAmount}
-/>
+<MoneyCell {symbol} amount={formattedAmount} onUpdate={updateAmount} />