+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## 1.1.0
 
-- Added settings panel.
+- Added a new settings panel.
+- Fixed an issue where the select menu was not updating correctly.
+- Added support for creating new income streams.
+- Added support for creating new categories.